Fix Tap the Vein plant snapshot typing
