+++ src/lib/lenses/lineage.ts
@@ -1,35 +1,8 @@
 import { db } from '@/lib/db'
-
-export type LensGoalTraceSource = 'live' | 'attach_snapshot' | 'plant_snapshot'
-
-export type LensGoalTraceNode = {
-  id: string
-  title: string
-  domain: string
-  cadence: string
-  status: string
-}
-
-export type LensGoalTrace = {
-  source: LensGoalTraceSource
-  capturedAt: string
-  goal: LensGoalTraceNode
-  parentChain: LensGoalTraceNode[]
-}
+import { isLensGoalTrace, type LensGoalTrace, type LensGoalTraceNode, type LensGoalTraceSource } from '@/lib/lenses/lineage-types'
 
 type GoalRow = LensGoalTraceNode & {
   parentGoalId: string | null
-}
-
-export function isLensGoalTrace(value: unknown): value is LensGoalTrace {
-  return Boolean(
-    value &&
-      typeof value === 'object' &&
-      'source' in value &&
-      'capturedAt' in value &&
-      'goal' in value &&
-      'parentChain' in value,
-  )
 }
 
 async function loadGoalChain(goalId: string, playerId: string): Promise<GoalRow[] | null> {
